Refactor startup data collection into modular, focused scripts
