+++ digital_health_startups_complete.csv
@@ -1,0 +1,110 @@
+url,company_name,products,location,has_company_name,has_products,has_location
+https://www.alfa-ai.com,ALFA AI GmbH,ALFA,,Yes,Yes,No
+https://www.avayl.tech/,AVAYL GmbH,AvaKit Sensor; Avayl Connect; AvaWave,,Yes,Yes,No
+https://www.acalta.de,Acalta GmbH,Acalta Gesundheits-App,,Yes,Yes,No
+https://www.accurx.com,Accurx | The simple way to communicate about patient care,,,Yes,No,No
+https://www.actimi.com,Actimi GmbH,Actimi 3D Sensor; Actimi Core,Stuttgart,Yes,Yes,Yes
+https://www.ada.com,Ada,,,Yes,No,No
+https://www.emmora.de,Ahorn AG,Rehappy App,,Yes,Yes,No
+https://www.aporize.com/,Aporize,Aporize Care,,Yes,Yes,No
+https://www.arztlena.com/,Artificy GmbH,Arzt Lena,,Yes,Yes,No
+https://shop.getnutrio.com/,Aurora Life Sciene GmbH,Nutrio,,Yes,Yes,No
+https://www.auta.health/,Auta Health UG,Auta Health App,Essen,Yes,Yes,Yes
+https://www.avimedical.com/avi-impact,Avi Medical Operations GmbH,Avi Medical App,Berlin,Yes,Yes,Yes
+https://de.becureglobal.com/,BECURE GmbH,BeCure Health Assistant,,Yes,Yes,No
+https://bellehealth.co/de/,Belle Health GmbH,Belle Health App,Bonn,Yes,Yes,Yes
+https://brea.app/,Brea Health GmbH,Brea App,,Yes,Yes,No
+https://breathment.com/,Breathment GmbH,Breathment App,,Yes,Yes,No
+https://www.careanimations.de/,CAREANIMATIONS GmbH,CareAnimations,,Yes,Yes,No
+https://de.caona.eu/,Caona Health GmbH,Caona Platform,,Yes,Yes,No
+https://www.caresyntax.com,Caresyntax,,,Yes,No,No
+https://sfs-healthcare.com,Change IT Solutions GmbH,SFS Platform,,Yes,Yes,No
+https://www.climedo.de/,Climedo Health GmbH,Climedo Platform,,Yes,Yes,No
+https://www.cliniserve.de/,Clinicserve GmbH,CliniServe Software,Rostock,Yes,Yes,Yes
+https://cogthera.de/#erfahren,Cogthera GmbH,Cogthera App,Bonn,Yes,Yes,Yes
+https://curecurve.de/elina-app/,CureCurve Medical AI GmbH,ELINA App,,Yes,Yes,No
+https://www.cynteract.com/de/rehabilitation,Cynteract GmbH,Cynteract Rehab System,,Yes,Yes,No
+https://www.healthmeapp.de/de/,Declareme GmbH,HealthMe App,,Yes,Yes,No
+https://denton-systems.de/,Denton Systems GmbH,Denton Solutions,,Yes,Yes,No
+https://www.skinly.de,Domain Details Page,,,Yes,No,No
+http://dopavision.com/,Dopavision GmbH,MyopiaCare,,Yes,Yes,No
+https://elixionmedical.com/,Elixion Medical,Elixion Platform,,Yes,Yes,No
+https://www.empident.de/,Empident GmbH,Empident Software,,Yes,Yes,No
+https://www.fitwhit.de,FitwHit & LABOR FÜR BIOMECHANIK der JLU-Gießen,FitWHit App,Friedberg,Yes,Yes,Yes
+https://www.floy.com/,Floy GmbH,Floy App,,Yes,Yes,No
+https://www.glaice.de/,GLACIE Health UG,Glaice Platform,,Yes,Yes,No
+https://gleea.de/,Gleea Educational Software GmbH,Gleea Software,"Klosterlechfeld
+Deutschland",Yes,Yes,Yes
+https://www.guidecare.de/,GuideCare GmbH,GuideCare Platform,,Yes,Yes,No
+https://www.apodienste.com/,Healthy Codes GmbH,APO-Dienste,,Yes,Yes,No
+https://www.help-app.de/,Help Mee Schmerztherapie GmbH,HELP App,,Yes,Yes,No
+https://home.informme.info/,InformMe GmbH,InformMe Platform,München,Yes,Yes,Yes
+https://www.inveox.com,Inveox,,München,Yes,No,Yes
+https://www.kenbi.eu,Kenbi - El ecológico,,,Yes,No,No
+https://www.kranushealth.com/de/therapien/haeufiger-harndrang,Kranus Health GmbH,Kranus Edera,,Yes,Yes,No
+https://www.kranushealth.com/de/therapien/inkontinenz,Kranus Health GmbH,Kranus Edera,,Yes,Yes,No
+https://www.kry.se,Kry,,,Yes,No,No
+https://www.lindera.de,LINDERA: Lösungen für digitale Sturzprävention und Bewegungsanalyse,,,Yes,No,No
+https://www.livi.co.uk,Livi,,,Yes,No,No
+https://www.echo.co.uk,LloydsDirect,,,Yes,No,No
+https://www.medadom.com,MEDADOM | Téléconsultation médicale en ligne 7j/7,,,Yes,No,No
+https://www.medgate.ch,Medgate | Immediate online doctor's appointment from home,,,Yes,No,No
+https://www.medipee.com,Medipee,,"Moers
+ 
+
+Seitenkarriere
+Über Uns
+Urinanalyse
+Wozu Urinanalyse
+Produkt
+Presse
+Kontakt
+Blog
+
+
+
+Gefördert Und Unterstützt Durch",Yes,No,Yes
+https://www.merantix.com,Merantix - Europe's leading AI group,,Berlin,Yes,No,Yes
+https://www.mika.health,Mika Health - Your digital companion for comprehensive cancer care,,,Yes,No,No
+https://www.mindable.health,Mindable Health,,,Yes,No,No
+https://www.smartpatient.eu,Pharma's Fully Integrated Digital Health Ecosystem,,Information Securityiso,Yes,No,Yes
+https://www.preventicus.com,Preventicus Heartbeats - Herzrhythmusstörungen messen,,Jena,Yes,No,Yes
+https://www.qare.fr,Qare - Téléconsultation médicale en ligne 7j/7,,,Yes,No,No
+https://www.teleclinic.com,TeleClinic,,,Yes,No,No
+https://www.thryve.health,Thryve | Secure Wearable Healthcare API Platform for Wearables & Health Apps,,,Yes,No,No
+https://www.tinnitracks.com,Tinnitus treatment with filtered music - Tinnitracks,,,Yes,No,No
+https://www.medwing.com,Top-Jobs Pflege & Medizin Plattform | MEDWING,,,Yes,No,No
+https://www.vitadock.com,VitaDock Online | VitaDock Online,,,Yes,No,No
+https://www.apheris.com,apheris AI GmbH,Apheris Compute Gateway,,Yes,Yes,No
+https://visioncheckout.com/,auvisus GmbH,Vision Checkout,,Yes,Yes,No
+https://www.biotx.ai/,biotx.ai GmbH,BioTX Platform,,Yes,Yes,No
+https://www.brainjo.de/,brainjo GmbH,Brainjo App,Regensburg,Yes,Yes,Yes
+https://www.comuny.de/,comuny GmbH,Comuny App,,Yes,Yes,No
+https://www.contextflow.com,contextflow – See beyond a single case,,,Yes,No,No
+https://deepeye.ai/,deepeye medical GmbH,DeepEye Platform,,Yes,Yes,No
+https://www.deepmentation.ai/,deepmentation UG,Deepmentation AI,,Yes,Yes,No
+https://www.derma2go.com/,derma2go Deutschland GmbH,derma2go App,,Yes,Yes,No
+https://www.dianovi.com/,dianovi GmbH (ehem. MySympto),Dianovi App,,Yes,Yes,No
+https://www.dpv-analytics.com/,dpv-analytics GmbH,DPV Analytics Platform,"Hamburg
+Rufen Sie Uns An",Yes,Yes,Yes
+http://www.ecovery.de/,eCovery GmbH,eCovery App,,Yes,Yes,No
+https://eye2you.ai/,eye2you,eye2you AI,,Yes,Yes,No
+https://www.felmo.de,felmo,,Berlin,Yes,No,Yes
+https://fyzo.de/assistant/,fyzo GmbH,fyzo Assistant; fyzo coach,,Yes,Yes,No
+https://www.gesund.de/app,gesund.de GmbH & Co. KG,gesund.de App,,Yes,Yes,No
+https://www.heartkinetics.com,heartkinetics.com,,,Yes,No,No
+https://www.heynanny.com/,heynannyly GmbH,heynannyly App,,Yes,Yes,No
+https://www.babylon.com,,,,No,No,No
+https://www.doctolib.de,,,,No,No,No
+https://www.doctolib.fr,,,,No,No,No
+https://www.kaia-health.com,,,,No,No,No
+https://www.medigene.com,,,,No,No,No
+https://www.sonormed.de,,,,No,No,No
+https://www.viomedo.de,,,,No,No,No
+https://www.xim.ai,,,,No,No,No
+https://www.zava.com,,,,No,No,No
+https://www.zavamed.com,,,,No,No,No
+https://incontalert.de/,inContAlert GmbH,inContAlert,,Yes,Yes,No
+https://www.mediteo.com,mediteo,,Heidelberg,Yes,No,Yes
+https://www.neosfer.de,neosfer,,,Yes,No,No
+https://www.samedi.de,samedi,,,Yes,No,No
